web/requirements/base.txt
Django==1.11.4

# DJANGO
django-debug-toolbar==1.8
django-extensions==1.8.1
djangorestframework==3.6.3
django-mptt==0.8.7
django_imagekit==4.0.1
django-redis==4.8.0
django-digg-paginator==0.1.3

# необходимо для django_imagekit
Pillow==4.2.1

# DJANGO END

# Jinja2
Jinja2==2.9.6

# django-environ for security issues
django-environ==0.4.3

# Celery
celery==4.1.0

# ELK Elasticsearch
elasticsearch==5.4.0
elasticsearch_dsl==5.3.0

# Django REST elasticsearch
django-rest-elasticsearch==0.3.2
requests==2.11.1
